chore: consolidate player-related scripts into `Player` folder
