+++ Assets/Scripts/CactusBushes/BushSeed.cs
@@ -1,5 +1,6 @@
 using System;
 using Cysharp.Threading.Tasks;
+using Player.Throwables;
 using SeedSnatcher.Target;
 using Unity.VisualScripting;
 using UnityEngine;
