+++ Assets/Scripts/Player/AttackAnimationScript.cs
