+++ Assets/Scripts/Player/AttackAnimationScript.cs.meta
