+++ Assets/Scripts/Player/Melee/TailWhip.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Attacks
+namespace Player.Melee
 {
     public class TailWhip : AbstractPlayerAttack
     {
