+++ Assets/Scripts/Player/PlayerPickup.cs
@@ -1,0 +1,21 @@
+using CactusBushes;
+using Player.Throwables;
+using UnityEngine;
+
+namespace Player
+{
+    public class PlayerPickup : MonoBehaviour
+    {
+        [SerializeField] private AmmoManager ammoManager;
+        [SerializeField] private Health health;
+    
+        private void OnCollisionEnter2D(Collision2D other)
+        {
+            if (other.gameObject.TryGetComponent<DroppableItem>(out var itemPickup))
+            {
+                itemPickup.OnCollect(health, ammoManager);
+                Destroy(itemPickup.gameObject);
+            }
+        }
+    }
+}
