+++ Assets/Scripts/Player/Throwables.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: e9af0a2f5501414fa688a1001b89f151
+timeCreated: 1745645444
